Max size for champion portrait (bug fix)
Currently, Wukong's portrait is loading the wrong image from ddragon. As seen here: http://champion.gg/champion/MonkeyKing (as of 6.1.1 patch). 

It's just loading a 1000x1000px green box and because there is no size limit on the site, it's covering the whole page. This change will prevent that from happening again.